requirements.txt
tensorflow
numpy
opencv-python-headless
streamlit
Pillow